+++ hutool-db/src/main/java/cn/hutool/db/handler/HandleHelper.java
@@ -109,7 +109,22 @@ public class HandleHelper {
 	 * @throws SQLException SQL执行异常
 	 */
 	public static Entity handleRow(int columnCount, ResultSetMetaData meta, ResultSet rs) throws SQLException {
-		return handleRow(Entity.create(), columnCount, meta, rs, true);
+		return handleRow(columnCount, meta, rs, false);
+	}
+	
+	/**
+	 * 处理单条数据
+	 * 
+	 * @param columnCount 列数
+	 * @param meta ResultSetMetaData
+	 * @param rs 数据集
+	 * @param caseInsensitive 是否大小写不敏感
+	 * @return 每一行的Entity
+	 * @throws SQLException SQL执行异常
+	 * @since 4.5.16
+	 */
+	public static Entity handleRow(int columnCount, ResultSetMetaData meta, ResultSet rs, boolean caseInsensitive) throws SQLException {
+		return handleRow(new Entity(null, caseInsensitive), columnCount, meta, rs, true);
 	}
 
 	/**
@@ -163,11 +178,26 @@ public class HandleHelper {
 	 * @throws SQLException SQL执行异常
 	 */
 	public static <T extends Collection<Entity>> T handleRs(ResultSet rs, T collection) throws SQLException {
+		return handleRs(rs, collection, false);
+	}
+	
+	/**
+	 * 处理多条数据
+	 * 
+	 * @param <T> 集合类型
+	 * @param rs 数据集
+	 * @param collection 数据集
+	 * @param caseInsensitive 是否大小写不敏感
+	 * @return Entity列表
+	 * @throws SQLException SQL执行异常
+	 * @since 4.5.16
+	 */
+	public static <T extends Collection<Entity>> T handleRs(ResultSet rs, T collection, boolean caseInsensitive) throws SQLException {
 		final ResultSetMetaData meta = rs.getMetaData();
 		final int columnCount = meta.getColumnCount();
 
 		while (rs.next()) {
-			collection.add(HandleHelper.handleRow(columnCount, meta, rs));
+			collection.add(HandleHelper.handleRow(columnCount, meta, rs, caseInsensitive));
 		}
 
 		return collection;