+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * 【core】         改进fnvHash避免负数（issue#IYDK6@Gitee）
 * 【core】         改进BeanCoper逻辑（pr#45@Gitee）
 * 【all】             实现必要序列化接口
+* 【db】            Entity增加可选忽略大小写（issue#IYGVW@Gitee）
+* 【core】         MapUtil增加renameKey方法（感谢@【帝都】宁静）
 
 ### Bug修复
 * 【poi】           修复sax中读取Excel普通单元格设置日期格式识别问题（issue#IYD0L@Gitee）